moving webviews telemetry into webviews
