+++ src/sharedInterfaces/telemetry.ts
@@ -42,57 +42,3 @@ export enum TelemetryActions {
     Close = "Close",
     SurverySubmit = "SurveySubmit",
 }
-
-export interface WebviewTelemetryActionEvent {
-    /**
-     * The view in which the event occurred.
-     */
-    telemetryView: TelemetryViews;
-    /**
-     * The action that was being performed when the event occurred.
-     */
-    telemetryAction: TelemetryActions;
-    /**
-     * Additional properties for the event.
-     */
-    additionalProps?: Record<string, string>;
-    /**
-     * Additional measurements for the event.
-     */
-    additionalMeasurements?: Record<string, number>;
-}
-
-export interface WebviewTelemetryErrorEvent {
-    /**
-     * The view in which the event occurred.
-     */
-    telemetryView: TelemetryViews;
-    /**
-     * The action that was being performed when the event occurred.
-     */
-    telemetryAction: TelemetryActions;
-    /**
-     * Error that occurred.
-     */
-    error: Error;
-    /**
-     * Whether to include the error message in the telemetry event. Defaults to false.
-     */
-    includeErrorMessage: boolean;
-    /**
-     * Error code for the error.
-     */
-    errorCode?: string;
-    /**
-     * Error type for the error.
-     */
-    errorType?: string;
-    /**
-     * Additional properties to include in the telemetry event.
-     */
-    additionalProps?: Record<string, string>;
-    /**
-     * Additional measurements to include in the telemetry event.
-     */
-    additionalMeasurements?: Record<string, number>;
-}
